+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* The `--generate-completions` option mistakenly still required a filename to
+  an action file.
+
 ## [1.0.1] - 2020-03-06
 
 ### Added